# openapi.yaml
openapi: 3.1.0
info:
  title: MiniFlow board API
  version: 0.1.0
  description: Implemented FastAPI contract. The backend is mounted at /api and stores boards, credentials, and tokens
    in a SQLAlchemy database (SQLite by default). WebSocket subscriptions remain in process memory. The frontend uses a random
    browser credential and bearer token for protected requests; there is no account-management UI. Public share-link
    reads require no token. Search, filters, and personal board-list ordering run in the frontend; board-list ordering
    is saved in that browser, not on this API.
servers:
  - url: /api
tags:
  - name: Auth
  - name: Boards
  - name: Columns
  - name: Tasks
paths:
  /auth/register:
    post:
      tags:
        - Auth
      operationId: register
      summary: Register a credential
      description: Used by the frontend to create a random browser credential. The password is stored as a salted
        PBKDF2-SHA256 hash. Registration requires a password of at least 8 characters. Credentials persist in the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '201':
          description: Credential registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredUser'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/token:
    post:
      tags:
        - Auth
      operationId: token
      summary: Exchange a credential for an opaque bearer token
      description: Invalid username or password returns 401. Tokens persist in the database across backend restarts.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /boards:
    get:
      tags:
        - Boards
      operationId: listBoards
      summary: List boards known to this browser
      description: Lists boards created or joined by the optional bearer identity. Without a valid token, returns
        an empty array. The frontend separately pins the public Demo Board.
      responses:
        '200':
          description: Boards available to this browser.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Board'
      security:
        - {}
        - bearerAuth: []
    post:
      tags:
        - Boards
      operationId: createBoard
      summary: Create a board from a built-in template
      description: Creates a board with template columns and no members or tasks. A valid optional bearer token
        associates it with that identity for listBoards. The creator joins separately before editing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoard'
      responses:
        '201':
          description: Created board.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '422':
          $ref: '#/components/responses/ValidationError'
      security:
        - {}
        - bearerAuth: []
  /boards/by-share/{shareId}:
    parameters:
      - $ref: '#/components/parameters/ShareId'
    get:
      tags:
        - Boards
      operationId: getBoardByShareId
      summary: Read a board using its shareable ID
      description: Public; the frontend maps 404 to null.
      responses:
        '200':
          description: Full board state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Board'
        '404':
          $ref: '#/components/responses/NotFound'
  /boards/{boardId}/viewer:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    get:
      tags:
        - Boards
      operationId: getViewer
      summary: Find this browser's membership on a board
      description: Returns the member for the optional bearer identity on this board, or JSON null. An unknown board
        returns 404.
      responses:
        '200':
          description: Current board member or null.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Member'
                  - type: 'null'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        - {}
        - bearerAuth: []
  /boards/{boardId}/members:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    post:
      tags:
        - Boards
      operationId: joinBoard
      summary: Join a board or update this browser's display name
      description: Requires a bearer token. Creates a membership for this identity on this board, or updates its
        display name while retaining the same member ID and joinedAt. Emits board.changed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinBoard'
      responses:
        '200':
          description: Current board member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - bearerAuth: []
  /boards/{boardId}/columns:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    post:
      tags:
        - Columns
      operationId: addColumn
      summary: Append a column
      security:
        - bearerAuth: []
      description: Position is the previous column count. A column named Done (case insensitive) becomes isDone
        only if the board has no Done column. Requires bearer token and membership on this board.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColumnName'
      responses:
        '201':
          description: Created column.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardColumn'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /boards/{boardId}/columns/{columnId}:
    parameters:
      - $ref: '#/components/parameters/BoardId'
      - $ref: '#/components/parameters/ColumnId'
    patch:
      tags:
        - Columns
      operationId: renameColumn
      summary: Rename a column
      security:
        - bearerAuth: []
      description: If renamed Done and no Done column exists, mark it isDone and set completedAt on its tasks that
        lack it. Renaming an existing Done column does not clear its designated isDone flag. Requires bearer token
        and membership on this board.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColumnName'
      responses:
        '204':
          description: Column renamed; no body.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags:
        - Columns
      operationId: deleteColumn
      summary: Delete a column and move its tasks to the first remaining column
      security:
        - bearerAuth: []
      description: Reject deletion of the last column. Append displaced tasks in their prior order; normalize column
        and task positions. Set or clear each moved task's completedAt according to the destination's isDone flag.
        Requires bearer token and membership on this board.
      responses:
        '204':
          description: Column deleted; no body.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /boards/{boardId}/columns/{columnId}/position:
    parameters:
      - $ref: '#/components/parameters/BoardId'
      - $ref: '#/components/parameters/ColumnId'
    put:
      tags:
        - Columns
      operationId: moveColumn
      summary: Move a column in one atomic write
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveColumn'
      responses:
        '204':
          description: Column moved; no body.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
      description: Requires bearer token and membership on this board.
  /boards/{boardId}/columns/{columnId}/tasks:
    parameters:
      - $ref: '#/components/parameters/BoardId'
      - $ref: '#/components/parameters/ColumnId'
    post:
      tags:
        - Tasks
      operationId: addTask
      summary: Append a task to a column
      security:
        - bearerAuth: []
      description: completedAt is set if the destination isDone; otherwise null. Requires bearer token and membership
        on this board.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskFields'
      responses:
        '201':
          description: Created task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /boards/{boardId}/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/BoardId'
      - $ref: '#/components/parameters/TaskId'
    put:
      tags:
        - Tasks
      operationId: updateTask
      summary: Replace a task's editable fields
      security:
        - bearerAuth: []
      description: Replace all TaskFields and update updatedAt. Position, columnId, completedAt, and creation time
        are controlled by the server. Requires bearer token and membership on this board.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskFields'
      responses:
        '204':
          description: Task updated; no body.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags:
        - Tasks
      operationId: deleteTask
      summary: Permanently delete a task
      security:
        - bearerAuth: []
      description: Normalize positions in the source column after deletion. Requires bearer token and membership
        on this board.
      responses:
        '204':
          description: Task deleted; no body.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /boards/{boardId}/tasks/{taskId}/position:
    parameters:
      - $ref: '#/components/parameters/BoardId'
      - $ref: '#/components/parameters/TaskId'
    put:
      tags:
        - Tasks
      operationId: moveTask
      summary: Move or reorder one task in one atomic write
      security:
        - bearerAuth: []
      description: Remove the task, insert it at toIndex in the destination (clamped to its valid insertion range),
        and renumber affected columns. Set completedAt on entry to the designated Done column and clear it on exit.
        Preserve completedAt when reordering inside Done. Requires bearer token and membership on this board.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveTask'
      responses:
        '204':
          description: Task moved; no body.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /boards/{boardId}/events:
    parameters:
      - $ref: '#/components/parameters/BoardId'
    get:
      tags:
        - Boards
      operationId: subscribeToBoard
      summary: Upgrade to a board WebSocket
      description: Public WebSocket upgrade for a board ID. After each member, column, or task mutation, the server
        sends board.changed with no board payload. Clients fetch the board after each frame and after reconnecting;
        the frontend polls every three seconds while disconnected. Subscriptions work within a single backend process
        only.
      x-websocket:
        server-message:
          $ref: '#/components/schemas/BoardChangedEvent'
        client-message: null
        unknown-board-close-code: 4404
      responses:
        '101':
          description: Switching Protocols to WebSocket.
        '403':
          description: Unknown board ID; WebSocket upgrade rejected before acceptance.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque token returned by POST /auth/token. A board membership is also required for column and
        task writes.
  parameters:
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: string
    ShareId:
      name: shareId
      in: path
      required: true
      schema:
        type: string
    ColumnId:
      name: columnId
      in: path
      required: true
      schema:
        type: string
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Missing or invalid bearer token, invalid credentials, or no membership on this board.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Board, column, or task not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Duplicate username or an operation that violates a board invariant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ValidationError:
      description: Invalid request body or field value.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required:
        - detail
      properties:
        detail:
          type: string
      description: FastAPI error responses use a string detail. Request validation returns the first validation
        message.
    TemplateId:
      type: string
      enum:
        - default
        - software
        - weekly
    Priority:
      oneOf:
        - type: string
          enum:
            - low
            - medium
            - high
        - type: 'null'
    Member:
      type: object
      additionalProperties: false
      required:
        - id
        - displayName
        - joinedAt
      properties:
        id:
          type: string
        displayName:
          type: string
          minLength: 1
        joinedAt:
          type: string
          format: date-time
    BoardColumn:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - position
        - isDone
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 1
        position:
          type: integer
          minimum: 0
        isDone:
          type: boolean
    TaskFields:
      type: object
      required:
        - title
        - description
        - priority
        - dueDate
        - assigneeId
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        dueDate:
          oneOf:
            - type: string
              format: date
            - type: 'null'
        assigneeId:
          description: Must be a member of this board when non-null.
          oneOf:
            - type: string
            - type: 'null'
    Task:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/TaskFields'
        - type: object
          required:
            - id
            - columnId
            - position
            - completedAt
            - createdAt
            - updatedAt
          properties:
            id:
              type: string
            columnId:
              type: string
            position:
              type: integer
              minimum: 0
            completedAt:
              oneOf:
                - type: string
                  format: date-time
                - type: 'null'
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    Board:
      type: object
      additionalProperties: false
      required:
        - id
        - shareId
        - name
        - template
        - createdAt
        - columns
        - tasks
        - members
      properties:
        id:
          type: string
        shareId:
          type: string
        name:
          type: string
          minLength: 1
        template:
          $ref: '#/components/schemas/TemplateId'
        createdAt:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: '#/components/schemas/BoardColumn'
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'
      description: Full board snapshot. The backend includes the public seeded Demo Board at shareId demo.
    CreateBoard:
      type: object
      additionalProperties: false
      required:
        - name
        - template
      properties:
        name:
          type: string
          minLength: 1
        template:
          $ref: '#/components/schemas/TemplateId'
    JoinBoard:
      type: object
      additionalProperties: false
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 1
    ColumnName:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
    MoveColumn:
      type: object
      additionalProperties: false
      required:
        - toIndex
      properties:
        toIndex:
          type: integer
          minimum: 0
    MoveTask:
      type: object
      additionalProperties: false
      required:
        - toColumnId
        - toIndex
      properties:
        toColumnId:
          type: string
        toIndex:
          type: integer
          minimum: 0
    BoardChangedEvent:
      type: object
      additionalProperties: false
      required:
        - type
      properties:
        type:
          type: string
          const: board.changed
    Credentials:
      type: object
      additionalProperties: false
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
          description: Registration requires at least 8 characters; token requests accept a string of at least 1
            character.
    RegisteredUser:
      type: object
      required:
        - username
      properties:
        username:
          type: string
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: bearer